pyproject.toml
[project]
name = "realtime"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "akazdayo", email = "82073147+akazdayo@users.noreply.github.com" }
]
dependencies = [
    "opencv-python>=4.11.0.86",
    "numpy>=2.2.4",
    "httpx>=0.28.1",
    "pillow>=11.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/realtime"]
